💥 Rename Ember specific scope option (#631)
This SDK compensates for Ember's testing UI by utilizing a `scope` option to dump the contents of
that scoped element into the document body. We semi-recently implemented a new type of `scope`
option directly in our API that allows scoping snapshots to specific elements in the DOM. With this
feature, the two scope options conflict, and this SDK cannot use the new element scoping feature
while the Ember testing UI scope feature exists.

Rather than rename the element scoping option for this SDK only, to keep all SDKs consistent we've
decided to rename the existing Ember SDK scope option to `emberTestingScope`. Unfortunately, that
makes this a breaking change for anybody relying on the old `scope` option behavior. When upgrading,
if not renamed, the new `scope` option will take effect and any cropping for that scoped element
will be done by our infrastructure and might have unexpected results.